:recycle: Move user milstones into its own component - Switch to using Apollo hooks - Remove dependency on Redux - Use withRouter to get current location - Didn't clean up the styling because this component needs to be redesigned. Not sure it is very useful at the moment.
