+++ app/javascript/pages/Dashboard/Milestones/milestoneQuery.gql
@@ -1,0 +1,6 @@
+query MilestoneQuery($after: Int, $before: Int) {
+  currentUser {
+    id
+    hours(after: $after, before: $before)
+  }
+}
